#54730 (Resources > Pool: Edit temp project for user): fix code flow comment
